[React] upgrade storybook to lastest version 5.3.18 (#958)
* upgrade storybook to lastest version 5.3.14

* add changelog

* fix mix of CSF

* npm install

* remove config console

* log url

* install @storybook/preset-create-react-app

* Add missing main.js

* update to latest storybook and associated configs. remove shared assets dependency.

* copy readme from root for now.

* remove config.js

* reorg of theming per instructions. still doesnt seem to be used.

* update storybook to 5.3.18

* stupid test removed

* update readme to indicate the latest version we are upgrading to.

* remove fsevents package which was added to debug osx `npm i` problems

* eslint

* put back shared assets and bad theme config

* test backstop on ci

* remove unneeded shared assets part of webpack and rebuild package-lock

* comment theme vars that were not being used before.

* last minute clean up

* package cleanup, get rid of nwb, and updates while we are at it

* add nwb back

* point to root README. remove babelrc. reinstall.

* use SharedAssets instead of relative paths. Remove copied typography.md

* linting
