--- react/.storybook/addons.js
@@ -1,6 +1,0 @@
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';
-import '@storybook/addon-viewport/register';
-import '@storybook/addon-storysource/register';
-import '@storybook/addon-a11y/register';
