+++ react/package.json
@@ -59,17 +59,18 @@
     "svg-sprite-loader": "^4.1.6"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-export-default-from": "^7.2.0",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@storybook/addon-a11y": "^5.2.3",
-    "@storybook/addon-actions": "^5.2.3",
-    "@storybook/addon-info": "^5.2.3",
-    "@storybook/addon-knobs": "^5.2.3",
-    "@storybook/addon-links": "^5.2.3",
-    "@storybook/addon-storysource": "^5.2.3",
-    "@storybook/addon-viewport": "^5.2.3",
-    "@storybook/addons": "^5.2.3",
-    "@storybook/react": "^5.2.3",
+    "@storybook/addon-a11y": "^5.3.18",
+    "@storybook/addon-actions": "^5.3.18",
+    "@storybook/addon-info": "^5.3.18",
+    "@storybook/addon-knobs": "^5.3.18",
+    "@storybook/addon-links": "^5.3.18",
+    "@storybook/addon-storysource": "^5.3.18",
+    "@storybook/addon-viewport": "^5.3.18",
+    "@storybook/addons": "^5.3.18",
+    "@storybook/react": "^5.3.18",
+    "@storybook/source-loader": "^5.3.18",
     "acorn": "^5.7.2",
     "acorn-jsx": "^4.1.1",
     "acorn-object-rest-spread": "^1.1.0",
@@ -77,9 +78,9 @@
     "babel-eslint": "^8.2.6",
     "babel-plugin-dynamic-import-node": "^2.2.0",
     "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-import-paths": "^1.0.0",
+    "babel-plugin-transform-import-paths": "^1.0.2",
     "cross-env": "^5.2.0",
-    "eslint": "^4.19.1",
+    "eslint": "^5.16.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-import-resolver-alias": "^1.1.1",
     "eslint-loader": "^2.1.0",
@@ -89,14 +90,14 @@
     "eslint-plugin-react-hooks": "^1.0.1",
     "github-markdown-css": "^3.0.1",
     "markdown-it": "^10.0.0",
-    "node-sass-chokidar": "^1.2.2",
-    "npm-run-all": "^4.1.2",
-    "nwb": "0.23.x",
+    "node-sass-chokidar": "^1.4.0",
+    "npm-run-all": "^4.1.5",
+    "nwb": "^0.24.5",
     "pre-push": "^0.1.1",
     "prop-types": "^15.7.1",
     "react": "16.8.6",
     "react-dom": "16.9.0",
-    "react-scripts": "^2.1.3"
+    "react-scripts": "^2.1.8"
   },
   "homepage": "",
   "license": "MIT",
