+++ react/src/components/organisms/Footer/Footer.stories.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, object, boolean, text } from '@storybook/addon-knobs';
-
 import stateSeal from 'SharedAssets/images/stateseal.png';
 import Footer from './index';
 import FooterDocs from './Footer.md';
