+++ react/src/components/organisms/Header/index.js
@@ -2,12 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import is from 'is';
 import classNames from 'classnames';
+import logo from 'SharedAssets/images/stateseal.png';
 import UtilityNav from '../UtilityNav';
 import MainNav from '../../molecules/MainNav';
 import HeaderSearch from '../../molecules/HeaderSearch';
 import SiteLogo from '../../atoms/media/SiteLogo';
-import logo from '../../../assets/images/stateseal.png';
-
 import './styles.css';
 
 class Header extends Component {
