+++ react/src/components/organisms/HeaderSlim/HeaderSlim.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, number, text } from '@storybook/addon-knobs';
 
-import logo from '../../../assets/images/stateseal.png';
+import logo from 'SharedAssets/images/stateseal.png';
 import HeaderSlim from '.';
 import { SiteLogo } from '../../../index';
